UTIL, add find_eos_delim and use it to simplify util_read_header (pending con)
Use memmem so we don't need '\0' ended buffer, we know most sizes anyway.

We now get a refbuf for free, so we attach it to the client already.

Util: use flags instead of the partial/entire mess, code looks cleaner that way

we'll need to port back the connection changes.

Signed-off-by: Niv Sardi <nsardi@smartjog.com>